feat: resolved linting issues
