+++ server/middleware/body-parser.ts
@@ -1,4 +1,4 @@
 import bodyParser from "body-parser"
 const handler = bodyParser.json()
 
-export default handler;
+export default handler
