+++ server/util/getGQLClient.ts
@@ -1,9 +1,7 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import {
-  CreateApolloClient,
-  KiboApolloClient,
-} from "@kibocommerce/graphql-client"
-import type { UserAuthTicket } from "@kibocommerce/graphql-client/dist/lib/AuthClient"
+/* eslint-disable @typescript-eslint/no-explicit-any */
+import { CreateApolloClient, KiboApolloClient } from "@kibocommerce/graphql-client"
+
 import type { KiboIncomingMessage, Response } from "../types/Api"
 
 function createGQLClient(req: KiboIncomingMessage): KiboApolloClient {
@@ -12,13 +10,14 @@ function createGQLClient(req: KiboIncomingMessage): KiboApolloClient {
     clientId: process.env.KIBO_CLIENT_ID || "",
     sharedSecret: process.env.KIBO_SHARED_SECRET || "",
     apiHost: process.env.KIBO_API_HOST || "",
+    authHost: "",
   }
   return CreateApolloClient({
     api: clientConfig,
     clientAuthHooks: {
       onTicketChange: () => {},
       onTicketRead: () => {
-        return {} as UserAuthTicket
+        return {} as any
       },
       onTicketRemove: () => {},
     },
