+++ tsconfig.json
@@ -1,5 +1,8 @@
 {
-  "extends": "./.nuxt/tsconfig.json"
+  "extends": "./.nuxt/tsconfig.json",
+  "compilerOptions": {
+    "skipLibCheck": true,
+    "types": ["@types/jest"],
+    "allowSyntheticDefaultImports": true
+  }
 }
-
-
